+++ tests/test_scripts/test_immv_imcp.py
@@ -11,6 +11,7 @@ import               tempfile
 
 import               pytest
 
+import numpy as np
 import nibabel as nib
 
 from   fsl.utils.tempdir import tempdir
@@ -269,7 +270,7 @@ def test_imcp_script_shouldPass(move=False):
                         for inf in infiles:
                             img     = nib.load(op.join(tindir, inf),
                                                mmap=False)
-                            imghash = hash(img.get_data().tobytes())
+                            imghash = hash(np.asanyarray(img.dataobj).tobytes())
                             img = None
                             imageHashes.append(imghash)